+++ tests/test_cli.py
@@ -3,10 +3,10 @@ import arrow
 from itertools import combinations
 from datetime import datetime, timedelta
 
-from dateutil.tz import tzlocal
 import pytest
 
 from watson import cli
+from watson.cli import local_tz_info
 
 
 # Not all ISO-8601 compliant strings are recognized by arrow.get(str)
@@ -173,7 +173,7 @@ def test_report_invalid_date(runner, watson, test_dt):
 @pytest.mark.parametrize('at_dt', VALID_TIMES_DATA)
 def test_stop_valid_time(runner, watson, mocker, at_dt):
     mocker.patch('arrow.arrow.dt_datetime', wraps=datetime)
-    start_dt = datetime(2019, 4, 10, 14, 0, 0, tzinfo=tzlocal())
+    start_dt = datetime(2019, 4, 10, 14, 0, 0, tzinfo=local_tz_info())
     arrow.arrow.dt_datetime.now.return_value = start_dt
     result = runner.invoke(cli.start, ['a-project'], obj=watson)
     assert result.exit_code == 0
@@ -190,7 +190,7 @@ def test_stop_valid_time(runner, watson, mocker, at_dt):
 def test_start_valid_time(runner, watson, mocker, at_dt):
     # Simulate a start date so that 'at_dt' is older than now().
     mocker.patch('arrow.arrow.dt_datetime', wraps=datetime)
-    start_dt = datetime(2019, 4, 10, 14, 0, 0, tzinfo=tzlocal())
+    start_dt = datetime(2019, 4, 10, 14, 0, 0, tzinfo=local_tz_info())
     arrow.arrow.dt_datetime.now.return_value = (start_dt + timedelta(hours=1))
     result = runner.invoke(cli.start, ['a-project', '--at', at_dt], obj=watson)
     assert result.exit_code == 0