+++ CHANGELOG.md
@@ -13,9 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python 3.8 support (#402)
 - Python 3.9 support (#402)
 
+### Changed
+
+- Upgrade to major arrow release 1.0.0 (#407)
+
 ### Removed
 
 - Python 2.7 support (#305).
+- Python 3.5 support (#407).
 
 ## [1.10.0] - 2020-07-03