+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `log` command output can now be filtered to exclude projects and tags via `--ignore-project` and `--ignore-tag` (#395)
 - Python 3.8 support (#402)
 - Python 3.9 support (#402)
+- Support for the TZ environment variable to specify the local time zone (#391)
 
 ### Changed